+++ tests/fuzz.rs
@@ -3,7 +3,7 @@
 
 use hvmc::{
   fuzz::*,
-  run::{Addr, Port, Strict, Tag},
+  run::{Addr, Heap, Port, Strict, Tag},
   trace,
 };
 
@@ -18,7 +18,7 @@ fn fuzz_var_link_link_var() {
   trace::set_hook();
   Fuzzer::default().fuzz(|fuzz| {
     unsafe { trace::_reset_traces() };
-    let heap = Net::init_heap(256);
+    let heap = Heap::new(256);
     let mut net = Net::new(&heap);
     let x = net.alloc();
     let y = net.alloc();
@@ -63,7 +63,7 @@ fn fuzz_pri_link_link_pri() {
     unsafe { trace::_reset_traces() };
     let p = Port::new(Tag::Ctr, 0, Addr::NULL);
     let q = Port::new(Tag::Ctr, 1, Addr::NULL);
-    let heap = Net::init_heap(256);
+    let heap = Heap::new(256);
     let mut net = Net::new(&heap);
     let x = net.alloc();
     let a = Port::new_var(x.clone());
@@ -80,7 +80,7 @@ fn fuzz_pri_link_link_pri() {
         n1.link_wire_port(b.wire(), q);
       });
     });
-    assert!(n0.rdex.len() == 1 || n1.rdex.len() == 1);
+    assert!(n0.redexes.len() == 1 || n1.redexes.len() == 1);
     for x in [a, b] {
       assert_eq!(x.addr().val().read(), Port::FREE.0);
     }
@@ -92,7 +92,7 @@ fn fuzz_pri_link_link_pri() {
 fn fuzz_var_link_link_pri() {
   assert!(cfg!(not(feature = "_fuzz_no_free")));
   trace::set_hook();
-  let heap = Net::init_heap(256);
+  let heap = Heap::new(256);
   Fuzzer::default().fuzz(|fuzz| {
     unsafe { trace::_reset_traces() };
     let mut net = Net::new(&heap);
@@ -133,7 +133,7 @@ fn fuzz_var_link_link_pri() {
 fn fuzz_var_link_link_link_var() {
   assert!(cfg!(feature = "_fuzz_no_free"));
   trace::set_hook();
-  let heap = Net::init_heap(256);
+  let heap = Heap::new(256);
   Fuzzer::default().fuzz(|fuzz| {
     unsafe { trace::_reset_traces() };
     let mut net = Net::new(&heap);
